+++ app/Controllers/MailLogController.php
@@ -99,6 +99,12 @@ class MailLogController extends ViewController
 	}
 
 	public function showResults(): Response {
+		if (!$this->showMailStats()) {
+			$this->flashbag->add('warning', "Results are disabled");
+			$this->initUrls();
+			return new RedirectResponse($this->homepageUrl);
+		}
+
 		// enable form rendering support
 		$this->twigFormView($this->request);
 
@@ -175,6 +181,12 @@ class MailLogController extends ViewController
 	}
 
 	public function showReports(string $field, string $mode = 'count'): Response {
+		if (!$this->showMailStats()) {
+			$this->flashbag->add('warning', "Reports are disabled");
+			$this->initUrls();
+			return new RedirectResponse($this->homepageUrl);
+		}
+
 		if (empty($field) || !in_array($field, MailLog::REPORT_FIELDS)) {
 			$this->flashbag->add('error', "Field '{$field}' does not exist");
 			$this->initUrls();
@@ -777,7 +789,7 @@ class MailLogController extends ViewController
 		return new Response($this->twig->render('search.twig', [
 			'qidform' => $qidform->createView(),
 			'filters' => $filters,
-			'stats'   => $this->showMailStats($service, $filters),
+			'stats'   => $this->getMailStats($service, $filters),
 			'searchform' => $searchform->createView(),
 			'runtime' => $this->getRuntime(),
 			'flashes' => $this->flashbag->all(),
@@ -875,15 +887,21 @@ class MailLogController extends ViewController
 		return $configs;
 	}
 
-	private function showMailStats(MailLogService $service, array $filters): array {
-		if (Config::get('show_mail_stats')) {
-			if ($this->getIsAdmin() || Config::get('show_user_mail_stats')) {
-				// has applyUserScope
-				return $service->showStats($filters);
-			}
+	private function showMailStats(): bool {
+		if (Config::get('show_mail_stats') &&
+			 ($this->getIsAdmin() || Config::get('show_user_mail_stats'))) {
+			 return true;
+		}
+		return false;
+	}
+
+	private function getMailStats(MailLogService $service, array $filters): array {
+		if ($this->showMailStats()) {
+			// has applyUserScope
+			return $service->showStats($filters);
 		}
 
-		return array();
+		return [];
 	}
 
 }